Refactor to use submodules
Also use better names for concurrent handshakes. And an `into_inner`
method on `ConcurrentHandshakeTls`.
